Move db method and test to separate files
