+++ q2_moshpit/eggnog/__init__.py
@@ -7,9 +7,8 @@
 # ----------------------------------------------------------------------------
 
 
-from ._method import (
-    eggnog_diamond_search, eggnog_annotate, fetch_eggnog_db, fetch_diamond_db
-)
+from ._method import eggnog_diamond_search, eggnog_annotate
+from ._dbs import fetch_eggnog_db, fetch_diamond_db
 
 __all__ = [
     'eggnog_diamond_search', 'eggnog_annotate', 'fetch_eggnog_db',
