+++ q2_moshpit/eggnog/_method.py
@@ -18,7 +18,6 @@ from q2_types_genomics.reference_db import (
     EggnogRefDirFmt, DiamondDatabaseDirFmt
 )
 from q2_types.feature_data import DNAFASTAFormat
-from .._utils import run_command
 from q2_types_genomics.feature_data import (
     OrthologAnnotationDirFmt, MAGSequencesDirFmt
 )
@@ -133,66 +132,3 @@ def _annotate_seed_orthologs_runner(seed_ortholog, eggnog_db, sample_label,
         cmds.append('--dbmem')
 
     subprocess.run(cmds, check=True)
-
-
-def fetch_eggnog_db() -> EggnogRefDirFmt:
-    """
-    Downloads eggnog reference database using the
-    `download_eggnog_data.py` script from eggNOG. Here, this
-    script downloads 3 files amounting to 47Gb in total.
-    """
-
-    # Initialize output objects
-    eggnog_db = EggnogRefDirFmt()
-
-    # Define command.
-    # Meaning of flags:
-    # y: Answer yest to all prompts thrown by download_eggnog_data.py
-    # D: Do not download the Diamond database
-    # data_dir: location where to save downloads
-    cmd = [
-        "download_eggnog_data.py", "-y", "-D",
-        "--data_dir", str(eggnog_db.path)
-    ]
-    run_command(cmd)
-
-    # Return objects
-    return eggnog_db
-
-
-def fetch_diamond_db() -> DiamondDatabaseDirFmt:
-    """
-    Downloads diamond reference database using the
-    `download_eggnog_data.py` script from eggNOG. Here, this
-    script downloads 1 file (8.6 Gb).
-    """
-
-    # Initialize output objects
-    diamond_db = DiamondDatabaseDirFmt()
-
-    # Download Diamond DB
-    print("Starting download...")
-    run_command([
-        f"cd {str(diamond_db)} && ",
-        "wget -e robots=off -O ref_db.dmnd.gz ",
-        'http://eggnogdb.embl.de/download/emapperdb-5.0.2/'
-        'eggnog_proteins.dmnd.gz'
-    ])
-
-    # Decompressing file
-    print("Download completed. \nDecompressing file...")
-    run_command([
-        "gunzip", "ref_db.dmnd.gz"
-    ])
-
-    # Let user know that the process is done.
-    # The actual copying wil be taken care of by qiime behind the
-    # scenes.
-    print(
-        "Decompression completed. \n"
-        "Copying file from temporary directory to final location "
-        "(this will take a few minutes)..."
-    )
-
-    # Return object
-    return diamond_db
